[EI-47] [api] rename: 파일명 변경
